// src/Feature/Navigation/client/GlobalHeader/components/NavUser/models.ts
//    Copyright 2020 EPAM Systems, Inc.
//
//    Licensed under the Apache License, Version 2.0 (the "License");
//    you may not use this file except in compliance with the License.
//    You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
//    Unless required by applicable law or agreed to in writing, software
//    distributed under the License is distributed on an "AS IS" BASIS,
//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//    See the License for the specific language governing permissions and
//    limitations under the License.

import { AuthProcessState } from 'Feature/Account/client/Integration/Authentication';
import * as Commerce from 'Foundation/Commerce/client';
import * as Jss from 'Foundation/ReactJss/client';

export interface NavUserProps {
  commerceUser: Commerce.CommerceUserModel;
  cartQuantity: number;
  returnUrl: string;
  authProcess: AuthProcessState;
  StartAuth: (email: string, password: string) => void;
}

export interface NavUserState extends Jss.SafePureComponentState {
  userFormVisible: boolean;
  email: string;
  formValid: boolean;
}